requirements.txt
asgiref==3.3.1
Djangı==3.1.5
gunicorn==20.0.4
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0